requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
filelock==3.13.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.4
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==2.1.5
mpmath==1.3.0
narwhals==2.2.0
networkx==3.2.1
numpy==1.26.3
packaging==25.0
pandas==2.3.2
pillow==11.0.0
protobuf==6.32.0
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
six==1.17.0
smmap==5.0.2
streamlit==1.49.0
sympy==1.13.3
tenacity==9.1.2
toml==0.10.2
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tornado==6.5.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
